% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AnnotationReader-class}
\alias{AnnotationReader-class}
\title{Abstract S4 class for AnnotationReader object}
\description{
Abstract S4 class for AnnotationReader object
}
\section{Slots}{

\describe{
\item{\code{file}}{character. - full path to annotation file}

\item{\code{annotation}}{data.frame. - dataframe contains annotation}
}}

