% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bggenes.R
\docType{data}
\name{bggenes}
\alias{bggenes}
\title{genes from rna-seq experiment on auxin treatment of Arabidopsis thaliana roots}
\format{A vector with 18039 elements.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE97258}
}
\usage{
bggenes
}
\description{
A vector containing the GeneIDs.
}
\keyword{datasets}
