% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bggenes_hum.R
\docType{data}
\name{bggenes_hum}
\alias{bggenes_hum}
\title{genes from rna-seq experiment on mRNA differential expression in LNCaP cells
expressing the wild-type androgen receptor (AR-WT) or the ligand-independent AR-V7 splice variant}
\format{A vector with 38238 elements.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE71334}
}
\usage{
bggenes_hum
}
\description{
A vector containing the GeneIDs.
}
\keyword{datasets}
