% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoldSpecTestObjMethods.R
\docType{methods}
\name{calcFSsignificance,FoldSpecTest-method}
\alias{calcFSsignificance,FoldSpecTest-method}
\title{Title}
\usage{
\S4method{calcFSsignificance}{FoldSpecTest}(object)
}
\arguments{
\item{FoldSpecTest}{- Object of FoldSpecTest class}
}
\description{
Title
}
