% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{calcFSsignificance}
\alias{calcFSsignificance}
\title{Run enrichment analysis}
\usage{
calcFSsignificance(object)
}
\arguments{
\item{object}{- Object of FoldSpecTest class}
}
\value{
object of FoldSpecTest class
}
\description{
Run enrichment analysis
}
