% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationReaderObjMethods.R
\docType{methods}
\name{convertToList,GAFReader-method}
\alias{convertToList,GAFReader-method}
\title{Convert GAF format type annotation to list contains GO term id's as keys and Gene ID's as values}
\usage{
\S4method{convertToList}{GAFReader}(object)
}
\arguments{
\item{GAFReader}{- object of GAFReader class}
}
\value{
- list with GO term id's as keys and Gene ID's as values
}
\description{
Convert GAF format type annotation to list contains GO term id's as keys and Gene ID's as values
}
\examples{
gaf_path <- system.file("extdata", "gene_association.tair.lzma", package = "FoldGO", mustWork = TRUE)
gaf <- GAFReader(file = gaf_path)
convertToList(gaf)
}
