% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{convertToList}
\alias{convertToList}
\title{Convert GAF format type annotation to list contains GO term id's as keys and Gene ID's as values}
\usage{
convertToList(object)
}
\arguments{
\item{object}{- object of GAFReader class}
}
\value{
- list with GO term id's as keys and Gene ID's as values
}
\description{
Convert GAF format type annotation to list contains GO term id's as keys and Gene ID's as values
}
\examples{
gaf_path <- system.file("extdata", "gene_association.tair.lzma",
                          package = "FoldGO", mustWork = TRUE)
gaf <- GAFReader(file = gaf_path, geneid_col = 10)
convertToList(gaf)
}
