% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degenes.R
\docType{data}
\name{degenes}
\alias{degenes}
\title{DEGs from rna-seq experiment on auxin treatment of Arabidopsis thaliana roots}
\format{A data frame with 789 rows and 4 variables:
\describe{
  \item{GeneID}{Gene identifier}
  \item{FC}{fold-change value}
  \item{pval}{p-value}
  \item{qval}{Benjamini-Yekutieli adjusted p-value}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE97258}
}
\usage{
degenes
}
\description{
A dataset containing the GeneIDs and corresponding fold-change values.
}
\keyword{datasets}
