% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneGroupsMethods.R
\docType{methods}
\name{divToGroups,GeneGroups-method}
\alias{divToGroups,GeneGroups-method}
\title{Title}
\usage{
\S4method{divToGroups}{GeneGroups}(object)
}
\arguments{
\item{GeneGroups}{- Object of GeneGroups class}
}
\description{
Title
}
