% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneGroupsMethods.R
\docType{methods}
\name{divToGroups}
\alias{divToGroups}
\alias{divToGroups,GeneGroups-method}
\title{Divide initial set of genes in to quantiles}
\usage{
divToGroups(object)

\S4method{divToGroups}{GeneGroups}(object)
}
\arguments{
\item{object}{- Object of GeneGroups class}
}
\value{
object of GeneGroups class
}
\description{
Divide initial set of genes in to quantiles
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneGroups}: Divide initial set of genes in to quantiles
}}

