% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_annotobj.R
\docType{data}
\name{down_annotobj}
\alias{down_annotobj}
\title{Annotation object compiled from lists of down-regulated genes
from rna-seq experiment on auxin treatment of Arabidopsis thaliana roots}
\format{An object of FuncAnnotGroupsTopGO class.}
\usage{
down_annotobj
}
\description{
Annotation object compiled from lists of down-regulated genes
from rna-seq experiment on auxin treatment of Arabidopsis thaliana roots
}
\keyword{datasets}
