% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_groups.R
\docType{data}
\name{down_groups}
\alias{down_groups}
\title{GeneGroups object compiled from down-regulated genes
from rna-seq experiment on auxin treatment of Arabidopsis thaliana roots}
\format{An object of GeneGroups class.}
\usage{
down_groups
}
\description{
GeneGroups object compiled from down-regulated genes
from rna-seq experiment on auxin treatment of Arabidopsis thaliana roots
}
\keyword{datasets}
