% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examdata.R
\docType{data}
\name{examdata_bg}
\alias{examdata_bg}
\alias{bggenes}
\alias{bggenes_hum}
\title{Background sets of genes used in examples}
\format{A vector containing the GeneIDs with 18039 and 38238 for A. thaliana and H. sapiens correspondingly}
\source{
\enumerate{
\item A. thaliana and auxin: \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE97258}
\item H. sapiens LNCap AR-V7: \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE71334}
}
}
\usage{
bggenes

bggenes_hum
}
\description{
We used genes from two datasets in examples:
\enumerate{
\item RNA-seq experiment on auxin treatment of Arabidopsis thaliana roots (degenes)
\item RNA-seq experiment on mRNA differential expression in LNCaP cells expressing
the wild-type androgen receptor (AR-WT) or the ligand-independent AR-V7 splice variant (degenes_hum)
}
}
\examples{
# load background genes from A. thaliana RNA-seq experiment
data("bggenes")
}
\keyword{datasets}
