% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examdata.R
\docType{data}
\name{examdata_degs}
\alias{examdata_degs}
\alias{degenes}
\alias{degenes_hum}
\title{Differential expressed genes used in examples}
\format{A dataframes with 4 variables and 789 and 2079 for A. thaliana and H. sapiens correspondingly,
where colnames are:
\describe{
  \item{GeneID}{Gene identifier}
  \item{FC}{fold-change value}
  \item{pval}{p-value}
  \item{qval}{Benjamini-Yekutieli adjusted p-value}
}}
\source{
\enumerate{
\item A. thaliana and auxin: \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE97258}
\item H. sapiens LNCap AR-V7: \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE71334}
}
}
\usage{
degenes

degenes_hum
}
\description{
We used two datasets in examples:
\enumerate{
\item RNA-seq experiment on auxin treatment of Arabidopsis thaliana roots (degenes)
\item RNA-seq experiment on mRNA differential expression in LNCaP cells expressing
the wild-type androgen receptor (AR-WT) or the ligand-independent AR-V7 splice variant (degenes_hum)
}
}
\keyword{datasets}
