% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examdata.R
\docType{data}
\name{examdata_objs}
\alias{examdata_objs}
\alias{up_groups}
\alias{down_groups}
\alias{up_annotobj}
\alias{down_annotobj}
\title{Precompiled objects of GeneGroups, FuncAnnotGroups
classes used in examples}
\format{\describe{
\item{up_groups}{object of GeneGroups class}
\item{down_groups}{object of GeneGroups class}
\item{up_annotobj}{object of FuncAnnotGroupsTopGO class}
\item{up_annotobj}{object of FuncAnnotGroupsTopGO class}
}}
\usage{
up_groups

down_groups

up_annotobj

down_annotobj
}
\description{
\describe{
\item{up_groups}{object of GeneGroups class compiled from up-regulated genes
from rna-seq experiment on auxin treatment of Arabidopsis thaliana roots}
\item{down_groups}{object of GeneGroups class compiled from down-regulated genes
from rna-seq experiment on auxin treatment of Arabidopsis thaliana roots}
\item{up_annotobj}{object of FuncAnnotGroups class compiled from lists of up-regulated genes
from rna-seq experiment on auxin treatment of Arabidopsis thaliana roots}
\item{up_annotobj}{object of FuncAnnotGroups class compiled from lists of down-regulated genes
from rna-seq experiment on auxin treatment of Arabidopsis thaliana roots}
}
}
\examples{
# load GeneGroups object with up-regulated genes from rna-seq experiment on auxin treatment
# of Arabidopsis thaliana roots
data("up_groups")
# load FuncAnnotGroups object compiled from lists of up-regulated genes
# from rna-seq experiment on auxin treatment of Arabidopsis thaliana roots
data("up_annotobj")
}
\keyword{datasets}
