% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoldSpecTestObjMethods.R
\docType{methods}
\name{findFSterms,FoldSpecTest-method}
\alias{findFSterms,FoldSpecTest-method}
\title{find fold-change-specific terms}
\usage{
\S4method{findFSterms}{FoldSpecTest}(object, fdrstep2 = NULL)
}
\arguments{
\item{fdrstep2}{- FDR threshold for 2 step of fold-specificty recognition procedure}

\item{FoldSpecTest}{- Object of FoldSpecTest class}
}
\value{
object of FoldSpecTest class
}
\description{
find fold-change-specific terms
}
