% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{findFSterms}
\alias{findFSterms}
\title{Find fold-change-specific terms}
\usage{
findFSterms(object, fdrstep2 = NULL)
}
\arguments{
\item{object}{- Object of FoldSpecTest class}

\item{fdrstep2}{- FDR threshold for 2 step of fold-specificty recognition procedure}
}
\value{
object of FoldSpecTest class
}
\description{
Find fold-change-specific terms
}
