% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationMethods.R
\name{fold_spec_chart}
\alias{fold_spec_chart}
\title{fold specificity rectangle plot}
\usage{
fold_spec_chart(data, interval_labels, x_text_size = 10)
}
\arguments{
\item{data}{dataframe contains GO term names and coordinates for rectangles both for up and down regulation}

\item{interval_labels}{vector of user defined names for non-overlaping intervals (e.g. c("weak response",...,"strong response"))}

\item{x_text_size}{size of text for x axis labels}
}
\value{
fold specificity rectangle plot as ggplot object
}
\description{
Create rectangle plot for fold specificity data representation
}
