% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationMethods.R
\name{fold_spec_chart_data}
\alias{fold_spec_chart_data}
\title{Data parser for fold-specificity rectangle plot}
\usage{
fold_spec_chart_data(fs_res_up, fs_res_down, nfs_res_up, nfs_res_down,
  wholeintname)
}
\arguments{
\item{fs_res_up}{- dataframe contains fold-specific GO terms and related data for up-regulation}

\item{fs_res_down}{- dataframe contains fold-specific GO terms and related data for down-regulation}

\item{nfs_res_up}{- dataframe contains not fold-specific GO terms and related data for up-regulation}

\item{nfs_res_down}{- dataframe contains not fold-specific GO terms and related data for down-regulation}

\item{wholeintname}{- name of the interval containing all differentially expressed genes}
}
\value{
input dataframe for create.fold.spec.chart function
}
\description{
Create input data for rectangle plot (fold_spec_chart function) using recognize_fs_terms function output as input
}
