% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoldSpecTestObj.R
\docType{class}
\name{FoldSpecTest-class}
\alias{FoldSpecTest-class}
\title{FoldSpecTest S4 class}
\description{
FoldSpecTest S4 class
}
\section{Slots}{

\describe{
\item{\code{fstable}}{data.frame. - dataframe with fold-specific GO terms data}

\item{\code{nfstable}}{data.frame. - dataframe with not fold-specific GO terms data}

\item{\code{result_table}}{data.frame. - table contains all GO terms and related data}

\item{\code{annotgroups}}{FuncAnnotGroups. - Annotaion object}

\item{\code{wholeintname}}{character. - name of the interval containing all differentially expressed genes}

\item{\code{padjmethod}}{character. - method for multiple testing correction (to see all possible methods print: p.adjust.methods)
Benjamini-Hochberg by default}

\item{\code{fdrstep1}}{numeric. - FDR threshold for 1 step of fold-specificty recognition procedure}

\item{\code{fdrstep2}}{numeric. - FDR threshold for 2 step of fold-specificty recognition procedure}

\item{\code{fisher_alternative}}{character. - indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less". You can specify just the initial letter. Only used in the 2 by 2 case.}
}}

