% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoldSpecTestObj.R
\name{FoldSpecTest}
\alias{FoldSpecTest}
\title{Constructor for FoldSpecTest S4 class}
\usage{
FoldSpecTest(annotgroups, ...)
}
\arguments{
\item{annotgroups}{- Annotaion object}

\item{...}{- Other parameters:
\itemize{
\item fdrstep1 - FDR threshold for 1 step of fold-specificty recognition procedure (1 by default)
\item fdrstep2 - FDR threshold for 2 step of fold-specificty recognition procedure (0.05 by default)
\item padjmethod - method for multiple testing correction (to see all possible methods print: p.adjust.methods)
                   Benjamini-Hochberg ("BH") by default
\item fisher_alternative - indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less".
                           You can specify just the initial letter. ("greater" by default)
}}
}
\description{
Constructor function that creates object of FoldSpecTest class.
             It takes object which is instance of subclass of AnnotGroups class (e.g. FuncAnnotGroupsTopGO class)
             as a minimal set of input parameters. For more details see Arguments section.
}
\examples{
FoldSpecTest(up_annotobj)
FoldSpecTest(up_annotobj, fdrstep1 = 0.2, fdrstep2 = 0.01, padjmethod = "BY")
}
