% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationObj.R
\docType{class}
\name{FuncAnnotGroups-class}
\alias{FuncAnnotGroups-class}
\title{Abstract S4 class for FuncAnnotGroups object}
\description{
Abstract S4 class for FuncAnnotGroups object
}
\section{Slots}{

\describe{
\item{\code{genegroups}}{GeneGroups. - object of GeneGroups class}

\item{\code{bggenes}}{character. - vector contains background set of genes}

\item{\code{resultlist}}{list. - list with filenames as keys and annotaton data frames as values}

\item{\code{padjmethod}}{character. - method for multiple testing correction (to see all possible methods print: p.adjust.methods)
Benjamini-Hochberg by default}

\item{\code{qitborder}}{numeric. - minimal number of genes annotated to a term (1 by default)}

\item{\code{wholeintname}}{character. - name of the DEG interval (initial set of genes)}
}}

