% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationObj.R
\docType{class}
\name{FuncAnnotGroupsTopGO-class}
\alias{FuncAnnotGroupsTopGO-class}
\title{S4 class for FuncAnnotGroupsTopGO object}
\description{
S4 class for FuncAnnotGroupsTopGO object
}
\section{Slots}{

\describe{
\item{\code{namespace}}{character. - character string specifing GO namespace ("BP", "MF" or "CC")}

\item{\code{genesannot}}{numeric. - minimal number of genes annotated to a term in the annotation}

\item{\code{algorithm}}{character. - from TopGO manual: character string specifing which algorithm to use. The algorithms are shown by the topGO whichAlgorithms() function.}

\item{\code{statistic}}{character. - from TopGO manual: character string specifing which test to use. The statistical tests are shown by the topGO whichTests() function.}

\item{\code{annot}}{function. - from TopGO manual: These functions are used to compile a list of GO terms such that each element in the list is a character vector containing all the gene identifiers that are mapped to the respective GO term.}

\item{\code{GO2genes}}{list. - from TopGO manual: named list of character vectors.  The list names are GO identifiers. For each GO the character vector contains the genes identifiers which are mapped to it. Only the most specific annotations are required.}

\item{\code{mapping}}{character. - from TopGO manual: character string specifieng the name of the Bioconductor package containing the gene mappings for a specific organism. For example: mapping = "org.Hs.eg.db".}

\item{\code{ID}}{character. - from TopGO manual: character string specifing the gene identifier to use. Currently only the following identifiers can be used: c("entrez", "genbank", "alias", "ensembl", "symbol",     "genename", "unigene")}
}}

