% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GAFReader-class}
\alias{GAFReader-class}
\title{S4 class for GAFReader object}
\description{
S4 class for GAFReader object
}
\section{Slots}{

\describe{
\item{\code{version}}{character. - version of GAF file}

\item{\code{info}}{character. - information from GAF file header}
}}

