% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{GAFReader}
\alias{GAFReader}
\title{Constructor for GAFReader S4 class}
\usage{
GAFReader(file, ...)
}
\arguments{
\item{file}{- full path to annotation file}

\item{...}{- Other parameters:
\itemize{
\item geneid_col - index of column with Gene ID (2 by default)
}}
}
\value{
- object of GAFReader class
}
\description{
Constructor function that creates object of GAFReader class.
             As a parameter it takes full path to file of GAF format.
}
\examples{
gaf_path <- system.file("extdata", "gene_association.tair.lzma",
                         package = "FoldGO", mustWork = TRUE)
gaf <- GAFReader(file = gaf_path, geneid_col = 10)
}
