% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneGroups.R
\docType{class}
\name{GeneGroups-class}
\alias{GeneGroups-class}
\title{S4 class for Gene Groups}
\description{
S4 class for Gene Groups
}
\section{Slots}{

\describe{
\item{\code{inputtable}}{data.frame. - dataframe contains initial set of genes gene ID's in the first row
and corresponding fold change values in the second row}

\item{\code{logfold}}{logical. - TRUE if fold values are presented in log scale, otherwise is FALSE}

\item{\code{quannumber}}{numeric. - number of quantiles (e.g. 2,3,4...)}

\item{\code{groups}}{list. - list of gene sets for each quatile and all combinations}

\item{\code{intnames}}{character. - vector of intervals names}

\item{\code{wholeintname}}{character. - name of the interval containing all differentially expressed genes}

\item{\code{regtype}}{character. - regulation type (up or down)}
}}

