% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneGroups.R
\name{GeneGroups}
\alias{GeneGroups}
\title{Constructor for GeneGroups S4 class}
\usage{
GeneGroups(inputtable, quannumber, ...)
}
\arguments{
\item{inputtable}{- dataframe contains initial set of genes gene ID's in the first row
and corresponding fold change values in the second row}

\item{quannumber}{- number of quantiles (e.g. 2,3,4...)}

\item{logfold}{- TRUE if fold values are presented in log scale, otherwise is FALSE (TRUE by default)}
}
\value{
- object of GeneGroups class
}
\description{
Constructor function that creates object of GeneGroups class.
             It takes dataframe with genes ID's and fold values, number of quantiles
             and logical variable which must set to TRUE if fold values are presented in logarithmic scale,
             otherwise it must be set to FALSE value (TRUE by default) as parameters.
}
\examples{
GeneGroups(degenes, 6)
GeneGroups(degenes, 10)

}
