% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationReaderObjMethods.R
\docType{methods}
\name{getAnnotation,AnnotationReader-method}
\alias{getAnnotation,AnnotationReader-method}
\title{Get annotation from GAF file}
\usage{
\S4method{getAnnotation}{AnnotationReader}(object)
}
\arguments{
\item{object}{- Object that is instance of subclass of AnnotationReader class}
}
\value{
- dataframe contains annotation
}
\description{
Get annotation from GAF file
}
