% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AnnotationReaderObjMethods.R
\docType{methods}
\name{getAnnotation}
\alias{getAnnotation}
\alias{getAnnotation,AnnotationReader-method}
\title{Get annotation derived from annotation file}
\usage{
getAnnotation(object)

\S4method{getAnnotation}{AnnotationReader}(object)
}
\arguments{
\item{object}{- Object that is instance of subclass of AnnotationReader class}
}
\value{
- dataframe contains annotation
}
\description{
Get annotation derived from annotation file
}
\section{Methods (by class)}{
\itemize{
\item \code{AnnotationReader}: Get annotation from GAF file
}}

\examples{
gaf_path <- system.file("extdata", "gene_association.tair.lzma",
                          package = "FoldGO", mustWork = TRUE)
gaf <- GAFReader(file = gaf_path, geneid_col = 10)
getAnnotation(gaf)
}
