% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getFStable}
\alias{getFStable}
\title{Get dataframe with fold-change-specific terms}
\usage{
getFStable(object)
}
\arguments{
\item{object}{- Object of FoldSpecTest class}
}
\value{
- dataframe with fold-specific GO terms data
}
\description{
Get dataframe with fold-change-specific terms
}
\examples{
fs_up <- FoldSpecTest(up_annotobj)
getFStable(fs_up)
}
