% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneGroupsMethods.R
\docType{methods}
\name{getGroups,GeneGroups-method}
\alias{getGroups,GeneGroups-method}
\title{Title}
\usage{
\S4method{getGroups}{GeneGroups}(object)
}
\arguments{
\item{GeneGroups}{- Object of GeneGroups class}
}
\value{
- list of gene sets for each quatile and all combinations
}
\description{
Title
}
\examples{
gene_groups <- GeneGroups(degenes, 6)
getGroups(gene_groups)
}
