% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getGroups}
\alias{getGroups}
\title{Get list of gene sets for each quatile and all combinations}
\usage{
getGroups(object)
}
\arguments{
\item{object}{- Object of GeneGroups class}
}
\value{
- list of gene sets for each quatile and all combinations
}
\description{
Get list of gene sets for each quatile and all combinations
}
\examples{
gene_groups <- GeneGroups(degenes, 6)
getGroups(gene_groups)
}
