% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneGroupsMethods.R
\docType{methods}
\name{getIntNames,GeneGroups-method}
\alias{getIntNames,GeneGroups-method}
\title{Title}
\usage{
\S4method{getIntNames}{GeneGroups}(object)
}
\arguments{
\item{GeneGroups}{- Object of GeneGroups class}
}
\value{
- vector of intervals names
}
\description{
Title
}
\examples{
gene_groups <- GeneGroups(degenes, 6)
getIntNames(gene_groups)
}
