% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneGroupsMethods.R
\docType{methods}
\name{getIntNames}
\alias{getIntNames}
\alias{getIntNames,GeneGroups-method}
\title{Get vector of intervals names}
\usage{
getIntNames(object)

\S4method{getIntNames}{GeneGroups}(object)
}
\arguments{
\item{object}{- Object of GeneGroups class}
}
\value{
- vector of intervals names
}
\description{
Get vector of intervals names
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneGroups}: Get vector of intervals names
}}

\examples{
gene_groups <- GeneGroups(degenes, 6)
getIntNames(gene_groups)
}
