% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getNFStable}
\alias{getNFStable}
\title{Get dataframe with not fold-change-specific terms}
\usage{
getNFStable(object)
}
\arguments{
\item{object}{- Object of FoldSpecTest class}
}
\value{
- dataframe with not fold-specific GO terms data
}
\description{
Get dataframe with not fold-change-specific terms
}
\examples{
fs_up <- FoldSpecTest(up_annotobj)
getNFStable(fs_up)
}
