% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneGroupsMethods.R
\docType{methods}
\name{getQuanNumber,GeneGroups-method}
\alias{getQuanNumber,GeneGroups-method}
\title{Title}
\usage{
\S4method{getQuanNumber}{GeneGroups}(object)
}
\arguments{
\item{GeneGroups}{- Object of GeneGroups class}
}
\value{
- number of quantiles
}
\description{
Title
}
\examples{
gene_groups <- GeneGroups(degenes, 6)
getQuanNumber(gene_groups)
}
