% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getQuanNumber}
\alias{getQuanNumber}
\title{Get number of quantiles}
\usage{
getQuanNumber(object)
}
\arguments{
\item{object}{- Object of GeneGroups class}
}
\value{
- number of quantiles
}
\description{
Get number of quantiles
}
\examples{
gene_groups <- GeneGroups(degenes, 6)
getQuanNumber(gene_groups)
}
