% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneGroupsMethods.R
\docType{methods}
\name{getRegType,GeneGroups-method}
\alias{getRegType,GeneGroups-method}
\title{Title}
\usage{
\S4method{getRegType}{GeneGroups}(object)
}
\arguments{
\item{GeneGroups}{- Object of GeneGroups class}
}
\value{
- regulation type (up or down)
}
\description{
Title
}
\examples{
gene_groups <- GeneGroups(degenes, 6)
getRegType(gene_groups)
}
