% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getRegType}
\alias{getRegType}
\title{Get regulation type}
\usage{
getRegType(object)
}
\arguments{
\item{object}{- Object of GeneGroups class}
}
\value{
- regulation type (up or down)
}
\description{
Get regulation type
}
\examples{
gene_groups <- GeneGroups(degenes, 6)
getRegType(gene_groups)
}
