% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getResultList}
\alias{getResultList}
\title{Accessor method for list of functional annotation result tables}
\usage{
getResultList(object)
}
\arguments{
\item{object}{- object of FuncAnnotGroups class}
}
\value{
- list with filenames as keys and annotaton data frames as values
}
\description{
Accessor method for list of functional annotation result tables
}
\examples{
library(topGO)
gaf_path <- system.file("extdata", "gene_association.tair.lzma",
                         package = "FoldGO", mustWork = TRUE)
gaf <- GAFReader(file = gaf_path)
gaf_list <- convertToList(gaf)
annotobj <- FuncAnnotGroupsTopGO(up_groups,"BP", GO2genes = gaf_list,
                                 annot = topGO::annFUN.GO2genes,
                                 bggenes = bggenes, padjmethod = "BH",
                                 qitborder = 10, genesannot = 1)
getResultList(annotobj)
}
