% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getResultTable}
\alias{getResultTable}
\title{Get dataframe with both fold-change-specific
and not fold-change-specific terms}
\usage{
getResultTable(object)
}
\arguments{
\item{object}{- Object of FoldSpecTest class}
}
\value{
- table contains all GO terms and related data
}
\description{
Get dataframe with both fold-change-specific
and not fold-change-specific terms
}
\examples{
fs_up <- FoldSpecTest(up_annotobj)
getResultTable(fs_up)
}
