% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationReaderObjMethods.R
\docType{methods}
\name{getVersion,GAFReader-method}
\alias{getVersion,GAFReader-method}
\title{Get version of GAF file}
\usage{
\S4method{getVersion}{GAFReader}(object)
}
\arguments{
\item{object}{- Object of GAFReader class}
}
\value{
- version of GAF file
}
\description{
Get version of GAF file
}
