% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AnnotationReaderObjMethods.R
\docType{methods}
\name{getVersion}
\alias{getVersion}
\alias{getVersion,GAFReader-method}
\title{Get version of instance}
\usage{
getVersion(object)

\S4method{getVersion}{GAFReader}(object)
}
\arguments{
\item{object}{- Object of S4 class}
}
\value{
- version of instance
}
\description{
Get version of instance
}
\section{Methods (by class)}{
\itemize{
\item \code{GAFReader}: Get version of GAF file
}}

\examples{
gaf_path <- system.file("extdata", "gene_association.tair.lzma",
                         package = "FoldGO", mustWork = TRUE)
gaf <- GAFReader(file = gaf_path, geneid_col = 10)
getVersion(gaf)
}
