% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneGroupsMethods.R
\docType{methods}
\name{getWholeIntName,GeneGroups-method}
\alias{getWholeIntName,GeneGroups-method}
\title{Get name of the interval containing all differentially expressed genes}
\usage{
\S4method{getWholeIntName}{GeneGroups}(object)
}
\arguments{
\item{object}{- Object of GeneGroups class}
}
\value{
- name of the interval containing all differentially expressed genes
}
\description{
Get name of the interval containing all differentially expressed genes
}
