% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/FoldSpecTestObjMethods.R,
%   R/GeneGroupsMethods.R
\docType{methods}
\name{getWholeIntName}
\alias{getWholeIntName}
\alias{getWholeIntName,FoldSpecTest-method}
\alias{getWholeIntName,GeneGroups-method}
\title{Get name of the interval containing all differentially expressed genes}
\usage{
getWholeIntName(object)

\S4method{getWholeIntName}{FoldSpecTest}(object)

\S4method{getWholeIntName}{GeneGroups}(object)
}
\arguments{
\item{object}{- Object of GeneGroups or FoldSpecTest class}
}
\value{
- name of the interval containing all differentially expressed genes
}
\description{
Get name of the interval containing all differentially expressed genes
}
\section{Methods (by class)}{
\itemize{
\item \code{FoldSpecTest}: Get name of largest fold-change interval (DEG interval)

\item \code{GeneGroups}: Get name of the interval
containing all differentially expressed genes
}}

\examples{
gene_groups <- GeneGroups(degenes, 6)
getWholeIntName(gene_groups)
fs_up <- FoldSpecTest(up_annotobj)
getWholeIntName(fs_up)
}
