% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationMethods.R
\docType{methods}
\name{plot,FoldSpecTest,ANY-method}
\alias{plot,FoldSpecTest,ANY-method}
\title{Fold-change specific GO Profile chart plotting}
\usage{
\S4method{plot}{FoldSpecTest,ANY}(x, y, x_text_size = 10)
}
\arguments{
\item{x}{- object of S4 FoldSpecTest class with up-regulated genes}

\item{y}{- object of S4 FoldSpecTest class with down-regulated genes}

\item{x_text_size}{- x axis labels size}
}
\value{
- Fold-change specific GO Profile plot
}
\description{
Fold-change specific GO Profile chart plotting
}
\examples{
# calculate fold-specificity test for up-regulated genes
up_fs <- FoldSpecTest(up_annotobj)
# calculate fold-specificity test for down-regulated genes
down_fs <- FoldSpecTest(down_annotobj)
plot(up_fs, down_fs)
}
