% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationReaderObjMethods.R
\docType{methods}
\name{read,GAFReader-method}
\alias{read,GAFReader-method}
\title{Title}
\usage{
\S4method{read}{GAFReader}(object)
}
\arguments{
\item{GAFReader}{- Object of GAFReader class}
}
\value{
- object of GAFReader class
}
\description{
Title
}
