% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AnnotationReaderObjMethods.R
\docType{methods}
\name{read}
\alias{read}
\alias{read,GAFReader-method}
\title{Read annotation file}
\usage{
read(object)

\S4method{read}{GAFReader}(object)
}
\arguments{
\item{object}{- object that is instance of subclass of AnnotationReader class (e.g. GAFReader class)}
}
\value{
- instance of subclass of AnnotationReader class
}
\description{
Read annotation file
}
\section{Methods (by class)}{
\itemize{
\item \code{GAFReader}: Read GAF file
}}

