% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationObjMethods.R
\docType{methods}
\name{runFuncAnnotTest,FuncAnnotGroupsTopGO-method}
\alias{runFuncAnnotTest,FuncAnnotGroupsTopGO-method}
\title{Functional annotation test}
\usage{
\S4method{runFuncAnnotTest}{FuncAnnotGroupsTopGO}(object)
}
\arguments{
\item{FuncAnnotGroupsTopGO}{- object of FuncAnnotGroupsTopGO class}
}
\description{
Functional annotation test
}
