% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{runFuncAnnotTest}
\alias{runFuncAnnotTest}
\title{Functional annotation test}
\usage{
runFuncAnnotTest(object)
}
\arguments{
\item{object}{- object of FuncAnnotGroupsTopGO class}
}
\value{
- object of FuncAnnotGroupsTopGO class
}
\description{
Functional annotation test
}
