% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/up_annotobj.R
\docType{data}
\name{up_annotobj}
\alias{up_annotobj}
\title{Annotation object compiled from lists of up-regulated genes
from rna-seq experiment on auxin treatment of Arabidopsis thaliana roots}
\format{An object of FuncAnnotGroupsTopGO class.}
\usage{
up_annotobj
}
\description{
Annotation object compiled from lists of up-regulated genes
from rna-seq experiment on auxin treatment of Arabidopsis thaliana roots
}
\keyword{datasets}
