% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/up_groups.R
\docType{data}
\name{up_groups}
\alias{up_groups}
\title{GeneGroups object compiled from up-regulated genes
from rna-seq experiment on auxin treatment of Arabidopsis thaliana roots}
\format{An object of GeneGroups class.}
\usage{
up_groups
}
\description{
GeneGroups object compiled from up-regulated genes
from rna-seq experiment on auxin treatment of Arabidopsis thaliana roots
}
\keyword{datasets}
