% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/FoldSpecTestObjMethods.R
\docType{methods}
\name{calcFSsignificance}
\alias{calcFSsignificance}
\alias{calcFSsignificance,FoldSpecTest-method}
\title{Run enrichment analysis}
\usage{
calcFSsignificance(object)

\S4method{calcFSsignificance}{FoldSpecTest}(object)
}
\arguments{
\item{object}{- Object of FoldSpecTest class}
}
\value{
object of FoldSpecTest class
}
\description{
Run enrichment analysis
}
\section{Methods (by class)}{
\itemize{
\item \code{FoldSpecTest}: Run enrichment analysis
}}

