% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AnnotationReaderObjMethods.R
\docType{methods}
\name{convertToList}
\alias{convertToList}
\alias{convertToList,GAFReader-method}
\title{Convert annotation to list contains GO term id's as keys and Gene ID's as values}
\usage{
convertToList(object)

\S4method{convertToList}{GAFReader}(object)
}
\arguments{
\item{object}{- object that is instance of subclass of AnnotationReader class (e.g. GAFReader class)}
}
\value{
- list with GO term id's as keys and Gene ID's as values
}
\description{
Convert annotation to list contains GO term id's as keys and Gene ID's as values
}
\section{Methods (by class)}{
\itemize{
\item \code{GAFReader}: Convert GAF format type annotation to list contains GO term id's as keys and Gene ID's as values
}}

\examples{
gaf_path <- system.file("extdata", "gene_association.tair.lzma",
                          package = "FoldGO", mustWork = TRUE)
gaf <- GAFReader(file = gaf_path, geneid_col = 10)
convertToList(gaf)
}
