% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{divToGroups}
\alias{divToGroups}
\title{Divide initial set of genes in to quantiles}
\usage{
divToGroups(object)
}
\arguments{
\item{object}{- Object of GeneGroups class}
}
\value{
object of GeneGroups class
}
\description{
Divide initial set of genes in to quantiles
}
