% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationReaderObjMethods.R
\docType{methods}
\name{getAnnotation,AnnotationReader-method}
\alias{getAnnotation,AnnotationReader-method}
\title{Title}
\usage{
\S4method{getAnnotation}{AnnotationReader}(object)
}
\arguments{
\item{AnnotationReader}{- Object that is instance of subclass of AnnotationReader class}
}
\value{
- dataframe contains annotation
}
\description{
Title
}
\examples{
gaf_path <- system.file("extdata", "gene_association.tair", package = "FoldGO", mustWork = TRUE)
gaf <- GAFReader(file = gafpath)
getAnnotation(gaf)
}
