% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getAnnotation}
\alias{getAnnotation}
\title{Get annotation from GAF file}
\usage{
getAnnotation(object)
}
\arguments{
\item{object}{- Object that is instance of subclass of AnnotationReader class}
}
\value{
- dataframe contains annotation
}
\description{
Get annotation from GAF file
}
\examples{
gaf_path <- system.file("extdata", "gene_association.tair.lzma",
                          package = "FoldGO", mustWork = TRUE)
gaf <- GAFReader(file = gaf_path)
getAnnotation(gaf)
}
