% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoldSpecTestObjMethods.R
\docType{methods}
\name{getFStable,FoldSpecTest-method}
\alias{getFStable,FoldSpecTest-method}
\title{get dataframe with fold-change-specific terms}
\usage{
\S4method{getFStable}{FoldSpecTest}(object)
}
\arguments{
\item{FoldSpecTest}{- Object of FoldSpecTest class}
}
\value{
- dataframe with fold-specific GO terms data
}
\description{
get dataframe with fold-change-specific terms
}
\examples{
fs_up <- FoldSpecTest(up_annotobj)
getFStable(fs_up)
}
