% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneGroupsMethods.R
\docType{methods}
\name{getGroups}
\alias{getGroups}
\alias{getGroups,GeneGroups-method}
\title{Get list of gene sets for each quatile and all combinations}
\usage{
getGroups(object)

\S4method{getGroups}{GeneGroups}(object)
}
\arguments{
\item{object}{- Object of GeneGroups class}
}
\value{
- list of gene sets for each quatile and all combinations
}
\description{
Get list of gene sets for each quatile and all combinations
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneGroups}: Get list of gene sets for each
quatile and all combinations
}}

\examples{
gene_groups <- GeneGroups(degenes, 6)
getGroups(gene_groups)
}
