% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getIntNames}
\alias{getIntNames}
\title{Get vector of intervals names}
\usage{
getIntNames(object)
}
\arguments{
\item{object}{- Object of GeneGroups class}
}
\value{
- vector of intervals names
}
\description{
Get vector of intervals names
}
\examples{
gene_groups <- GeneGroups(degenes, 6)
getIntNames(gene_groups)
}
