% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoldSpecTestObjMethods.R
\docType{methods}
\name{getNFStable,FoldSpecTest-method}
\alias{getNFStable,FoldSpecTest-method}
\title{get dataframe with not fold-change-specific terms}
\usage{
\S4method{getNFStable}{FoldSpecTest}(object)
}
\arguments{
\item{FoldSpecTest}{- Object of FoldSpecTest class}
}
\value{
- dataframe with not fold-specific GO terms data
}
\description{
get dataframe with not fold-change-specific terms
}
\examples{
fs_up <- FoldSpecTest(up_annotobj)
getNFStable(fs_up)
}
