% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/FoldSpecTestObjMethods.R
\docType{methods}
\name{getNFStable}
\alias{getNFStable}
\alias{getNFStable,FoldSpecTest-method}
\title{Get dataframe with not fold-change-specific terms}
\usage{
getNFStable(object)

\S4method{getNFStable}{FoldSpecTest}(object)
}
\arguments{
\item{object}{- Object of FoldSpecTest class}
}
\value{
- dataframe with not fold-specific GO terms data
}
\description{
Get dataframe with not fold-change-specific terms
}
\section{Methods (by class)}{
\itemize{
\item \code{FoldSpecTest}: Get dataframe with not fold-change-specific terms
}}

\examples{
fs_up <- FoldSpecTest(up_annotobj)
getNFStable(fs_up)
}
