% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneGroupsMethods.R
\docType{methods}
\name{getQuanNumber}
\alias{getQuanNumber}
\alias{getQuanNumber,GeneGroups-method}
\title{Get number of quantiles}
\usage{
getQuanNumber(object)

\S4method{getQuanNumber}{GeneGroups}(object)
}
\arguments{
\item{object}{- Object of GeneGroups class}
}
\value{
- number of quantiles
}
\description{
Get number of quantiles
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneGroups}: Get number of quantiles
}}

\examples{
gene_groups <- GeneGroups(degenes, 6)
getQuanNumber(gene_groups)
}
