% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneGroupsMethods.R
\docType{methods}
\name{getRegType}
\alias{getRegType}
\alias{getRegType,GeneGroups-method}
\title{Get regulation type}
\usage{
getRegType(object)

\S4method{getRegType}{GeneGroups}(object)
}
\arguments{
\item{object}{- Object of GeneGroups class}
}
\value{
- regulation type (up or down)
}
\description{
Get regulation type
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneGroups}: Get regulation type
}}

\examples{
gene_groups <- GeneGroups(degenes, 6)
getRegType(gene_groups)
}
