% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationObjMethods.R
\docType{methods}
\name{getResultList,FuncAnnotGroups-method}
\alias{getResultList,FuncAnnotGroups-method}
\title{Accessor method for list of functional annotation result tables}
\usage{
\S4method{getResultList}{FuncAnnotGroups}(object)
}
\arguments{
\item{FuncAnnotGroups}{- object of FuncAnnotGroups class}
}
\value{
- list with filenames as keys and annotaton data frames as values
}
\description{
Accessor method for list of functional annotation result tables
}
\examples{
gaf_path <- system.file("extdata", "gene_association.tair", package = "FsgorS4package", mustWork = TRUE)
gaf <- GAFReader(file = gafpath)
gaf_list <- convertToList(gaf)
annotobj <- FuncAnnotGroupsTopGO(up_groups,"BP", gene2GO = gaf_list, annot = topGO::annFUN.GO2genes, bggenes = bggenes, padjmethod = "BH", qitborder = 10, genesannot = 1)
getResultList(annotobj)
}
