% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getResultList}
\alias{getResultList}
\title{Accessor method for list of functional annotation result tables}
\usage{
getResultList(object)
}
\arguments{
\item{object}{- object of FuncAnnotGroups class}
}
\value{
- list with filenames as keys and annotaton data frames as values
}
\description{
Accessor method for list of functional annotation result tables
}
\examples{
getResultList(up_annotobj)
}
