% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AnnotationObjMethods.R
\docType{methods}
\name{getResultList}
\alias{getResultList}
\alias{getResultList,FuncAnnotGroups-method}
\title{Accessor method for list of functional annotation result tables}
\usage{
getResultList(object)

\S4method{getResultList}{FuncAnnotGroups}(object)
}
\arguments{
\item{object}{- object of FuncAnnotGroups class}
}
\value{
- list with filenames as keys and annotaton data frames as values
}
\description{
Accessor method for list of functional annotation result tables
}
\section{Methods (by class)}{
\itemize{
\item \code{FuncAnnotGroups}: Accessor method for list of functional annotation result tables
}}

\examples{
getResultList(up_annotobj)
}
