% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoldSpecTestObjMethods.R
\docType{methods}
\name{getResultTable,FoldSpecTest-method}
\alias{getResultTable,FoldSpecTest-method}
\title{Title}
\usage{
\S4method{getResultTable}{FoldSpecTest}(object)
}
\arguments{
\item{FoldSpecTest}{- Object of FoldSpecTest class}
}
\value{
- table contains all GO terms and related data
}
\description{
Title
}
\examples{
fs_up <- FoldSpecTest(up_groups)
getResultTable(fs_up)
}
