% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/FoldSpecTestObjMethods.R
\docType{methods}
\name{getResultTable}
\alias{getResultTable}
\alias{getResultTable,FoldSpecTest-method}
\title{Get dataframe with both fold-change-specific
and not fold-change-specific terms}
\usage{
getResultTable(object)

\S4method{getResultTable}{FoldSpecTest}(object)
}
\arguments{
\item{object}{- Object of FoldSpecTest class}
}
\value{
- table contains all GO terms and related data
}
\description{
Get dataframe with both fold-change-specific
and not fold-change-specific terms
}
\section{Methods (by class)}{
\itemize{
\item \code{FoldSpecTest}: Get dataframe with both fold-change-specific
and not fold-change-specific terms
}}

\examples{
fs_up <- FoldSpecTest(up_annotobj)
getResultTable(fs_up)
}
