% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationReaderObjMethods.R
\docType{methods}
\name{getVersion,GAFReader-method}
\alias{getVersion,GAFReader-method}
\title{Title}
\usage{
\S4method{getVersion}{GAFReader}(object)
}
\arguments{
\item{GAFReader}{- Object of GAFReader class}
}
\value{
- version of GAF file
}
\description{
Title
}
\examples{
gaf_path <- system.file("extdata", "gene_association.tair.lzma", package = "FoldGO", mustWork = TRUE)
gaf <- GAFReader(file = gaf_path)
getVersion(gaf)
}
