% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getVersion}
\alias{getVersion}
\title{Get version of GAF file}
\usage{
getVersion(object)
}
\arguments{
\item{object}{- Object of GAFReader class}
}
\value{
- version of GAF file
}
\description{
Get version of GAF file
}
\examples{
gaf_path <- system.file("extdata", "gene_association.tair.lzma",
                         package = "FoldGO", mustWork = TRUE)
gaf <- GAFReader(file = gaf_path)
getVersion(gaf)
}
