% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoldSpecTestObjMethods.R
\docType{methods}
\name{getWholeIntName,FoldSpecTest-method}
\alias{getWholeIntName,FoldSpecTest-method}
\title{Title}
\usage{
\S4method{getWholeIntName}{FoldSpecTest}(object)
}
\arguments{
\item{FoldSpecTest}{- Object of FoldSpecTest class}
}
\value{
- name of the interval containing all differentially expressed genes
}
\description{
Title
}
\examples{
fs_up <- FoldSpecTest(up_groups)
getWholeIntName(fs_up)
}
