% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoldSpecTestObjMethods.R
\docType{methods}
\name{getWholeIntName,FoldSpecTest-method}
\alias{getWholeIntName,FoldSpecTest-method}
\title{Get name of largest fold-change interval (DEG interval)}
\usage{
\S4method{getWholeIntName}{FoldSpecTest}(object)
}
\arguments{
\item{object}{- Object of FoldSpecTest class}
}
\value{
- name of the interval containing all differentially expressed genes
}
\description{
Get name of largest fold-change interval (DEG interval)
}
