% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneGroupsMethods.R
\docType{methods}
\name{getWholeIntName,GeneGroups-method}
\alias{getWholeIntName,GeneGroups-method}
\title{Title}
\usage{
\S4method{getWholeIntName}{GeneGroups}(object)
}
\arguments{
\item{GeneGroups}{- Object of GeneGroups class}
}
\value{
- name of the interval containing all differentially expressed genes
}
\description{
Title
}
\examples{
gene_groups <- GeneGroups(degenes, 6)
getWholeIntName(gene_groups)
}
