% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getWholeIntName}
\alias{getWholeIntName}
\title{Get name of the interval containing all differentially expressed genes}
\usage{
getWholeIntName(object)
}
\arguments{
\item{object}{- Object of GeneGroups or FoldSpecTest class}
}
\value{
- name of the interval containing all differentially expressed genes
}
\description{
Get name of the interval containing all differentially expressed genes
}
\examples{
gene_groups <- GeneGroups(degenes, 6)
getWholeIntName(gene_groups)
fs_up <- FoldSpecTest(up_annotobj)
getWholeIntName(fs_up)
}
