% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AnnotationObjMethods.R
\docType{methods}
\name{runFuncAnnotTest}
\alias{runFuncAnnotTest}
\alias{runFuncAnnotTest,FuncAnnotGroupsTopGO-method}
\title{Functional annotation test}
\usage{
runFuncAnnotTest(object)

\S4method{runFuncAnnotTest}{FuncAnnotGroupsTopGO}(object)
}
\arguments{
\item{object}{- object of FuncAnnotGroupsTopGO class}
}
\value{
- object of FuncAnnotGroupsTopGO class
}
\description{
Functional annotation test
}
\section{Methods (by class)}{
\itemize{
\item \code{FuncAnnotGroupsTopGO}: Functional annotation test
}}

